Fix diag OP bug on Windows Python3.8, cherry-pick from #28034
Fix diag OP bug on Windows Python3.8, remove the std::min
